requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==3.0.4
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
daphne==3.0.2
Django==5.0.2
django-environ==0.11.2
django-preventconcurrentlogins==0.8.2
django-rest-framework==0.1.0
django-user-agents==0.4.0
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.8
oauthlib==3.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.0.0
pytz==2024.1
redis==5.0.2
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
ua-parser==0.18.0
urllib3==2.2.1
user-agents==2.2.0
whitenoise==6.6.0
zope.interface==6.2
uvicorn==0.27.1
wsproto==1.2.0
psycopg2-binary==2.9.9
sentry-sdk==1.40.6